refatr router, add 422 for malformatted /slackCommand
